Project Initialization
Updated .gitignore to include the tests resource directory where drivers and feature files will be housed. Created Intellij Idea project. Updated pom.xml to include the dependencies we will need for the project. Added directory structure and example classes in so that the directory structure does not collapse on pushing to github.