SafariWebDriver Created & WebDriverManager Changes
Created a concrete class for the Safari Web Driver. Added a new method within the WebDriverManager that all of the driver classes now overload. This method is for initializing the driver with a custom file path rather than just the default. Added the option into the factory to create a safari web driver.